refactor #18 ファイルの移動と不要なファイルの削除
