+++ .idea/runConfigurations/balanced_processing_time.xml
@@ -2,7 +2,7 @@
   <configuration default="false" name="balanced-processing-time" type="Application" factoryName="Application">
     <option name="ALTERNATIVE_JRE_PATH" value="corretto-11" />
     <option name="ALTERNATIVE_JRE_PATH_ENABLED" value="true" />
-    <option name="MAIN_CLASS_NAME" value="org.eclipse.paho.sample.mqttv5app.ShareSample" />
+    <option name="MAIN_CLASS_NAME" value="org.eclipse.paho.sample.mqttv5app.sample.ShareSample" />
     <module name="org.eclipse.paho.sample.mqttv5app" />
     <option name="PROGRAM_PARAMETERS" value="-t 10 -s 3 -i 10 --ping-interval 1000 --process-time 25,25,25 --size 100" />
     <method v="2">
