--- org.eclipse.paho.sample.mqttv5app/src/main/java/org/eclipse/paho/sample/mqttv5app/Sample.java
@@ -1,9 +1,0 @@
-package org.eclipse.paho.sample.mqttv5app;
-
-
-
-public class Sample {
-    public static void main(String[] args) {
-        System.out.println("Hello World!");
-    }
-}
